feat(config): 新增龙猫 ini 实现和各种底层实现优化

+++ App/Configuration/Impl/JsonFileProvider.cs
@@ -1,13 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Text.Json;
 using System.Text.Json.Nodes;
 using PCL.Core.IO;
 
 namespace PCL.Core.App.Configuration.Impl;
 
-public class JsonFileProvider : CommonFileProvider
+/// <summary>
+/// 提供 JSON 格式的键值文件读写。
+/// </summary>
+public class JsonFileProvider : CommonFileProvider, IEnumerableKeyProvider
 {
     private readonly JsonObject _rootElement;
 
@@ -24,24 +28,47 @@ public class JsonFileProvider : CommonFileProvider
 
     public JsonFileProvider(string path) : base(path)
     {
-        JsonNode? parseResult;
         try
         {
-            using var stream = new FileStream(FilePath, FileMode.Open, FileAccess.Read, FileShare.Read);
-            parseResult = JsonNode.Parse(stream, documentOptions: _DocumentOptions);
+            if (File.Exists(path))
+            {
+                using var stream = new FileStream(FilePath, FileMode.Open, FileAccess.Read, FileShare.ReadWrite);
+                var parseResult = JsonNode.Parse(stream, documentOptions: _DocumentOptions);
+                if (parseResult is not JsonObject root)
+                    throw new FileInitException(path,
+                        $"Invalid root element type: {parseResult?.GetValueKind().ToString() ?? "Empty"}");
+                _rootElement = root;
+            }
+            else
+            {
+                using var stream = new FileStream(FilePath, FileMode.CreateNew, FileAccess.Write, FileShare.ReadWrite);
+                _rootElement = new JsonObject();
+                JsonSerializer.Serialize(stream, _rootElement, _SerializerOptions);
+            }
         }
-        catch (Exception ex) { throw new FileInitException(path, "Failed to read JSON file", ex); }
-        if (parseResult is not JsonObject rootElement) throw new FileInitException(path,
-            $"Invalid root element type: {parseResult?.GetValueKind().ToString() ?? "Empty"}");
-        _rootElement = rootElement;
+        catch (Exception ex)
+        {
+            if (ex is FileInitException) throw;
+            throw new FileInitException(path, "Failed to read JSON file", ex);
+        }
     }
 
     public override T Get<T>(string key)
     {
         var result = _rootElement[key];
-        return result == null
-            ? throw new KeyNotFoundException($"Not found: '{key}'")
-            : result.Deserialize<T>()!;
+        if (result == null) throw new KeyNotFoundException($"Not found: '{key}'");
+        try
+        {
+            var r = result.Deserialize<T>();
+            return r ?? throw new NullReferenceException(GetNullMessage());
+        }
+        catch (JsonException)
+        {
+            if (typeof(T) == typeof(string)) return (T)(object)result.ToString();
+            var jsonStr = result.Deserialize<string>()!;
+            return JsonSerializer.Deserialize<T>(jsonStr, _SerializerOptions) ?? throw new NullReferenceException(GetNullMessage());
+        }
+        string GetNullMessage() => $"Deserialized value is null: '{key}'";
     }
 
     public override void Set<T>(string key, T value)
@@ -61,9 +88,11 @@ public class JsonFileProvider : CommonFileProvider
 
     public override void Sync()
     {
-        File.Copy(FilePath, FilePath + ".bak", true);
+        if (File.Exists(FilePath)) File.Copy(FilePath, FilePath + ".bak", true);
         var stream = new FileStream(FilePath, FileMode.Create, FileAccess.Write, FileShare.Read);
         var writer = new Utf8JsonWriter(stream);
         _rootElement.WriteTo(writer, _SerializerOptions);
     }
+
+    public IEnumerable<string> Keys => _rootElement.Select(pair => pair.Key);
 }
